% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_overlap.R
\name{code_overlap}
\alias{code_overlap}
\title{code_overlap}
\usage{
code_overlap(connection = NULL, files = NULL)
}
\arguments{
\item{connection}{.rqda project file}

\item{files}{vector of files to restrict search to}
}
\description{
Adjacency Matrix at text level
}
\keyword{Cloud}
\keyword{Network}
\keyword{RQDA}
\keyword{Shiny}
