% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeIncase.R
\name{codeIncase}
\alias{codeIncase}
\title{codeIncase}
\usage{
codeIncase(case, code = NULL, output = "df", files = NULL,
  connection = NULL)
}
\arguments{
\item{case}{vector of "cases"}

\item{code}{vector of "codes" to search within the "cases"}

\item{output}{output format df=data frame ; HTML = HTML file}

\item{files}{vector of files to restrict search to}

\item{connection}{.rqda project file}
}
\description{
Find overlap in highlighted text between "codes" and "cases"
}
\keyword{Cloud}
\keyword{Network}
\keyword{RQDA}
\keyword{Shiny}
