% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_adjacency.R
\name{doc_adjacency}
\alias{doc_adjacency}
\title{doc_adjacency}
\usage{
doc_adjacency(connection = NULL, type = "product", include = "all",
  case = NULL, code = NULL, files = NULL)
}
\arguments{
\item{connection}{.rqda project file}

\item{type}{"product" or "unit". Where overlaps are multiple should they be multiplied ("product") or summed ("unit")}

\item{files}{vector of files to restrict search to}
}
\description{
Adjacency Matrix at document level
}
\keyword{Cloud}
\keyword{Network}
\keyword{RQDA}
\keyword{Shiny}
