% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQDAnetwork.R
\name{RQDAnetwork}
\alias{RQDAnetwork}
\title{RQDAnetwork}
\usage{
RQDAnetwork(adj, group = TRUE)
}
\arguments{
\item{adj}{adjacency matrix}

\item{group}{logical. Indicating whether to attempt to identify groups}
}
\description{
Build network analysis
}
\keyword{Cloud}
\keyword{Network}
\keyword{RQDA}
\keyword{Shiny}
