% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeIncase.R
\name{codeIncase}
\alias{codeIncase}
\title{codecase}
\usage{
codeIncase(case, code = NULL, files = NULL, connection = NULL,
  output = "df")
}
\arguments{
\item{case}{vector of "cases" to include as search domain}

\item{code}{vector of "codes" to search within the "cases"}

\item{files}{vector of files to include as search domain}

\item{connection}{.rqda project file}

\item{output}{output format df=data frame ; HTML = HTML file}
}
\description{
Find overlap in highlighted text between "codes" and "cases"
}
\keyword{Case}
\keyword{Cloud}
\keyword{Code}
\keyword{Network}
\keyword{Overlaps}
\keyword{RQDA}
\keyword{Shiny}
