% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQDAPlus.R
\name{RQDAPlus}
\alias{RQDAPlus}
\title{RQDAPlus}
\usage{
RQDAPlus(connection)
}
\arguments{
\item{connection}{.rqda project file}
}
\description{
Shiny interface to additional features using an RQDA project file
}
\keyword{Cloud}
\keyword{Network}
\keyword{RQDA}
\keyword{Shiny}
